atl: Replace "dim" argument with "crpix" in atlWcspx
svn path=/trunk/; revision=26739

+++ libraries/atl/atl.h
@@ -71,7 +71,7 @@ void atlPtftl( AstFitsChan *, const char *, int, const char *, int * );
 void atlPtftr( AstFitsChan *, const char *, float, const char *, int * );
 void atlPtfts( AstFitsChan *, const char *, const char *, const char *,  int * );
 void atlTolut( AstMapping *, double, double, double, const char *, AstMapping **, int * );
-void atlWcspx( AstKeyMap *, AstKeyMap *, int[3], double, double, AstFrameSet **, int * );
+void atlWcspx( AstKeyMap *, AstKeyMap *, double[3], double, double, AstFrameSet **, int * );
 
 
 #endif
